requirements.txt
streamlit
langchain
python-dotenv
pypdf
faiss-cpu
langchain-community
pytesseract
pdf2image
Pillow
beautifulsoup4
lxml
requests
sentence-transformers
langchain-groq